+++ src/Zatomic.AI.Providers/TogetherAI/TogetherAIChatRequest.cs
@@ -42,7 +42,7 @@ namespace Zatomic.AI.Providers.TogetherAI
 		public string SafetyModel { get; set; }
 
 		[JsonProperty("seed", NullValueHandling = NullValueHandling.Ignore)]
-		public long? Seed { get; set; }
+		public decimal? Seed { get; set; }
 
 		[JsonProperty("stop", NullValueHandling = NullValueHandling.Ignore)]
 		public List<string> Stop { get; set; }
